requirements.txt
streamlit
openai
python-dotenv
pydantic
langgraph>=0.2.0
langchain-core>=0.3.0
docker
